src/main/kotlin/com/maximum0/board/service/LikeService.kt
package com.maximum0.board.service

import com.maximum0.board.domain.Like
import com.maximum0.board.exception.PostNotFoundException
import com.maximum0.board.repository.LikeRepository
import com.maximum0.board.repository.PostRepository
import org.springframework.data.repository.findByIdOrNull
import org.springframework.stereotype.Service
import org.springframework.transaction.annotation.Transactional

@Service
@Transactional(readOnly = true)
class LikeService(
    private val likeRepository: LikeRepository,
    private val postRepository: PostRepository
) {
    fun createLike(postId: Long, createdBy: String): Long {
        val post = postRepository.findByIdOrNull(postId) ?: throw PostNotFoundException()
        return likeRepository.save(Like(post, createdBy)).id
    }
}
